requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
python-dotenv

